Ajustada arquitetura do projeto.
