+++ src/main/java/com/act/cooperativism/domain/repository/PautaRepository.java
@@ -1,9 +1,9 @@
-package com.act.cooperativism.repository;
+package com.act.cooperativism.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.act.cooperativism.model.Pauta;
+import com.act.cooperativism.domain.model.Pauta;
 
 @Repository
 public interface PautaRepository extends JpaRepository<Pauta, Long>{
