+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-conslole
+
 spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
